print route

+++ eclipse-workspace/java/src/java/Algorithm.java
@@ -1,11 +1,10 @@
-import java.util.ArrayList;
-
 public class Algorithm {
 	
 	private PathList pathlist;
 	private boolean[] nearEndpoint;
 	public final int nearKm=25;
 	private int pathnum;
+	final int INF = 100000;
 	//0번이 start_destination
 	//마지막 index가 final_destination
 	
@@ -14,17 +13,15 @@ public class Algorithm {
 		this.nearEndpoint= new boolean[JsonParser.MAX_CONTENTS];	//default : false
 		nearEndpoint();
 		this.pathnum = pathlist.getcount();
-		
-		for(int i=0; i<this.pathnum; i++)
-			this.pathlist.newArrayList(i);
 	}
 	
 	public void cal_distance(int num) {
 		double cal_latitude = Double.parseDouble(pathlist.getDestinationLat(num));
 		double cal_longtitude = Double.parseDouble(pathlist.getDestinationLon(num));
 		
+		pathlist.initialLength(num);
 		
-		for(int i=1; i<pathnum-1; i++) {
+		for(int i=num; i<pathnum-1; i++) {
 		
 			//자기자신은 제외
 			if(num == i) 
@@ -41,44 +38,54 @@ public class Algorithm {
 			dist = rad2deg(dist);
 			dist = dist * 60 * 1.1515;
 			
-			//kilometer
+			//kilometer	
 			dist = Math.round(dist * 1.609344 * 1000)/ 1000.0;	//셋째자리 까지 반올림하여 표시
 			
-			
+			System.out.println(num + " -> " + i + " : " + dist);
+			//pathlist.setLength(num, dist);
 			pathlist.setLength(num, dist, i);
-			
 		}
 		
 	}
 	
 	public void optimumRoute() {
 		for(int i=0; i<this.pathnum-2; i++) {
+			cal_distance(i);
+
 			int min_index = closeDestination(i);
+			
+			if(min_index == 0)
+				break;
 			
 			pathlist.addDestination(i+2, pathlist.getDestination(min_index));
 			pathlist.delete(min_index+2);
-			pathlist.printTravelRoute();	
+			pathlist.printTravelRoute();
+			
+			
 		}
 	}
 	
-	//가장 가까운 거리의 dest 찾기
 	public int closeDestination(int index) {
-		ArrayList<Double> temp;
+		double[] temp;
 		double min=0;
 		int min_index=0;
-		final int INF = 100000;
+		
 		
 		temp = pathlist.getLength(index);
 		min = INF;
-		for(int i=0; i<temp.size(); i++) {
+		
+		for(int i=0; i<pathnum -1; i++) {	
+			System.out.println("test" + i + ": " + temp[i]);
 			
-			
-			if(min > temp.get(i)) {
-				min = temp.get(i);
-				min_index= i;
+			if(temp[i] != 0 && pathlist.getVisit(i) == false) {
+				if(min > temp[i]) {
+					min = temp[i];
+					min_index= i;
+				}
 			}
 		}
-		min_index++;
+		
+		pathlist.setVisit(min_index);
 		System.out.println(min_index + ":" + min);
 		
 		return min_index;
